requirements.txt
absl-py==0.12.0
asgiref==3.3.2
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
docopt==0.6.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
idna==2.10
Js2Py==0.71
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MouseInfo==0.1.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
pipwin==0.5.1
protobuf==3.17.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/VishnuCharan/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
python-bidi==0.4.2
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
twilio==6.59.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.12.1
xhtml2pdf==0.2.4
